requirements.txt
fastapi
uvicorn
numpy
faiss-cpu
sentence-transformers
pydantic
pytest
httpx
